renamed the project to dicechanics
