--- DiceStatistics/__init__.py
@@ -1,6 +1,0 @@
-from DiceStatistics._Dice import Dice
-from DiceStatistics._Pool import Pool
-from DiceStatistics._BooleanDice import BooleanDice
-from DiceStatistics.interface import d, z, pool
-from DiceStatistics.standard import d4,d6,d8,d10,d12,d20,d100,z9
-from DiceStatistics._math import ceil, floor
