+++ dicechanics/__init__.py
@@ -1,0 +1,6 @@
+from dicechanics._Dice import Dice
+from dicechanics._Pool import Pool
+from dicechanics._BooleanDice import BooleanDice
+from dicechanics.interface import d, z, pool
+from dicechanics.standard import d4,d6,d8,d10,d12,d20,d100,z9
+from dicechanics._math import ceil, floor
