+++ tests/test_level2.py
@@ -1,6 +1,6 @@
 import unittest
 
-import DiceStatistics as tts
+import dicechanics as tts
 import dice_unittest
 
 class TestLevel2(dice_unittest.TestCase):
